+++ tests/contract_testcases/substrate/primitives/units.sol
@@ -1,6 +1,9 @@
 
         contract c {
             function test() public {
-                int32 x = 1 ether;
+                int64 x = 1 ether;
+                int64 wei = 1 sol;
+                // dot is not a unit, produce an error!
+                int64 dot = 1 dot;
             }
         }